Step R83: SharedTypeQueryRelationalTestBase adopted, by moving its base off Tier A (#56)
`failed` 181 -> 185, `total` 29214 -> 29218. A DELIBERATE RISE: 6 tests, 2 green, 4 red. FIXED
none, BROKEN exactly the 4, every one inside the new class. Compliance missing list 11 -> 10.
`test/` only, so `eng/measure.sh` and not the trim ratchet.

MOVED, NOT ADDED. `SharedTypeQueryRelationalTestBase` derives from `SharedTypeQueryTestBase`, which
this repository ran on Tier A. Adopting the relational base beside it would have run the shared base
on both tiers, which CLAUDE.md calls duplication rather than coverage. The class moved to Tier B
whole and its two inherited tests still pass there, so the move itself cost nothing.

THE 4 REDS ARE TWO MECHANISMS AND ONE WAS NOT KNOWN. Three of them fail with the same
`System.ArgumentException` inside EF's own `QueryFilterRewritingConvention`, while building THE
CLIENT'S MODEL, before any query runs: a `HasQueryFilter` whose body calls `FromSqlRaw` cannot be
rewritten for this client. That is not #60's runtime `FromSql` gap. It is a model-build failure, it
happens on the client alone, and no test named it before.

The fourth casts the test store to `RelationalTestStore` and then calls `SqlQueryRaw`. R77 would
remove the cast and not the red, because the call behind it needs raw SQL on a client with no
database. Second base measured against R77's premise, second to refuse it. R77 stays parked.

The handoff read this base as "3 tests, all reaching SqlQueryRaw/FromSqlRaw". The count was right
and the conclusion half right: three of the four never reach raw SQL at all. Reading which API a
test calls does not tell you where it fails.
