+++ test/known-failures.names.txt
@@ -53,6 +53,10 @@ InfoCarrier.Core.FunctionalTests.Sqlite.Query.PrimitiveCollectionsQuerySqliteInf
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.QueryNoClientEvalInfoCarrierTest.Doesnt_throw_when_from_sql_not_composed
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.QueryNoClientEvalInfoCarrierTest.Throws_when_from_sql_composed
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.QueryNoClientEvalInfoCarrierTest.Throws_when_orderby_multiple
+InfoCarrier.Core.FunctionalTests.Sqlite.Query.SharedTypeQueryInfoCarrierTest.Ad_hoc_query_for_default_shared_type_entity_type_throws
+InfoCarrier.Core.FunctionalTests.Sqlite.Query.SharedTypeQueryInfoCarrierTest.Ad_hoc_query_for_shared_type_entity_type_works
+InfoCarrier.Core.FunctionalTests.Sqlite.Query.SharedTypeQueryInfoCarrierTest.Can_use_shared_type_entity_type_in_query_filter_with_from_sql(async: False)
+InfoCarrier.Core.FunctionalTests.Sqlite.Query.SharedTypeQueryInfoCarrierTest.Can_use_shared_type_entity_type_in_query_filter_with_from_sql(async: True)
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.TPCGearsOfWarQueryInfoCarrierTest.Correlated_collection_after_distinct_3_levels_without_original_identifiers(async: False)
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.TPCGearsOfWarQueryInfoCarrierTest.Correlated_collection_after_distinct_3_levels_without_original_identifiers(async: True)
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.TPCGearsOfWarQueryInfoCarrierTest.Correlated_collection_order_by_constant_null_of_non_mapped_type(async: False)
